requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
whitenoise==6.5.0
dj-database-url==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
